Adds the AI channel ID configuration
Until now we were using only one channel for one purpose, AI announcements.
However, we're going to use the generalChannelId for the actual general channel now, and the aiChannelId for what was already doing.

This prevents confusion as it should've been called in the past something like "broadcastChannelId" or something along those lines.

General channel is going to be used by automated "Did You Know" announcements by the bot.